Condense comments
