+++ tools/tf2openapi/README.md
@@ -29,6 +29,16 @@ This project is motivated by the following uses of an OpenAPI specification:
 ### Project: TensorFlow-to-OpenAPI Transformer
 The outcome of this project is a TensorFlow-to-OpenAPI transformer which takes SavedModel protobuf binary files and generates an OpenAPI specification for prediction requests.
 
+#### Design Decisions
+There are numerous ways to format valid model input payloads to TFServing. Here are the formats this tool has chosen:
+* For batchable inputs with -1 in all 0-dimensions, uses the [row format](https://www.tensorflow.org/tfx/serving/api_rest#specifying_input_tensors_in_row_format)
+  * Only one named input tensor: uses `[val1, val2, etc.]` instead of equally valid `[{tensor: val1}, {tensor: val2}, ..]`
+  * Multiple named input tensors: uses `[{tensor1: val1, tensor2: val3, ..}, {tensor1: val2, tensor2: val4, ..}..]`
+* For non-batchable inputs (e.g. scalars) and batchable inputs without -1 in the 0th dimensions of all tensors to indicate that they are batchable, uses the [column format](https://www.tensorflow.org/tfx/serving/api_rest#specifying_input_tensors_in_column_format)
+  * Only one named input tensor: uses `val` instead of equally valid `{tensor: val}`
+  * Multiple named input tensors: uses `{tensor1: [val1, val2, ..], tensor2: [val3, val4, ..] ..}`
+
+Tf2OpenAPI generates a row format payload whenever possible because it's more intuitive for a user to construct.
 ## Caveats
 * There is a dependency on protobufs defined by TensorFlow, e.g. [tensorflow/core/protobuf](https://github.com/tensorflow/tensorflow/tree/master/tensorflow/core/protobuf). Specific protos must be compiled into Go using [protoc](https://github.com/golang/protobuf/tree/master/protoc-gen-go) in the order: tensorflow/core/lib/core/\*.proto, tensorflow/core/framework/\*.proto, tensorflow/core/protobuf/saver.proto, tensorflow/core/protobuf/meta_graph.proto, tensorflow/core/protobuf/saved_model.proto. See Makefile which will automate this.
