+++ tools/tf2openapi/types/tfsignaturedef.go
@@ -71,15 +71,23 @@ func NewTFMethod(key string, method string) (TFMethod, error) {
 	return tfMethod, nil
 }
 
+func (t *TFSignatureDef) Schema() *openapi3.Schema {
+	// https://www.tensorflow.org/tfx/serving/api_rest#specifying_input_tensors_in_row_format
+	if canHaveRowSchema(t.Inputs) {
+		return openapi3.NewObjectSchema().WithProperty("instances", t.rowSchema())
+	}
+
+	// https://www.tensorflow.org/tfx/serving/api_rest#specifying_input_tensors_in_column_format
+	return openapi3.NewObjectSchema().WithProperty("inputs", t.colSchema())
+}
+
 func canHaveRowSchema(t []TFTensor) bool {
 	for _, input := range t {
 		if input.Rank == -1 {
 			return false
 		}
-		// not batchable because either
-		// 1. tensor is a scalar or
-		// 2. model builder didn't follow the convention of having -1 in all 0th dim to
-		// indicate batchable inputs
+
+		// tensor is a scalar or tensor doesn't have -1 in 0th dim
 		if len(input.Shape) == 0 || input.Shape[0] != -1 {
 			return false
 		}
@@ -89,14 +97,11 @@ func canHaveRowSchema(t []TFTensor) bool {
 
 func (t *TFSignatureDef) rowSchema() *openapi3.Schema {
 	if len(t.Inputs) == 1 {
-		// only one named input tensor.
-		// although schema can be [{tensor: val1}, {tensor: val2}, ..]
-		// choose this schema of the form [val1, val2, etc.]
+		// e.g. [val1, val2, etc.]
 		singleTensorSchema := t.Inputs[0].RowSchema()
 		return openapi3.NewArraySchema().WithItems(singleTensorSchema)
 	}
-	// multiple named input tensors.
-	// schema is of the form [{tensor1: val1, tensor2: val3, ..}, {tensor1: val2, tensor2: val4, ..}..]
+	// e.g. [{tensor1: val1, tensor2: val3, ..}, {tensor1: val2, tensor2: val4, ..}..]
 	multiTensorSchema := openapi3.NewObjectSchema().WithProperties(make(map[string]*openapi3.Schema))
 	schema := openapi3.NewArraySchema().WithItems(multiTensorSchema)
 	for _, i := range t.Inputs {
@@ -108,16 +113,10 @@ func (t *TFSignatureDef) rowSchema() *openapi3.Schema {
 
 func (t *TFSignatureDef) colSchema() *openapi3.Schema {
 	if len(t.Inputs) == 1 {
-		// only one named input tensor.
-		// although schema can be {tensor name: val}
-		// choose the schema of the form val
-		// see description of FillTensorMapFromInputsMap in TFServing
-		// https://github.com/tensorflow/serving/blob/master/tensorflow_serving/util/json_tensor.cc
+		// e.g. val
 		return t.Inputs[0].ColSchema()
 	}
-
-	// multiple named input tensors.
-	// schema is of the form {tensor1: [val1, val2, ..], tensor2: [val3, val4, ..] ..}
+	// e.g. {tensor1: [val1, val2, ..], tensor2: [val3, val4, ..] ..}
 	schema := openapi3.NewObjectSchema().WithProperties(make(map[string]*openapi3.Schema))
 	for _, i := range t.Inputs {
 		schema.Properties[i.Name] = i.ColSchema().NewRef()
@@ -125,16 +124,4 @@ func (t *TFSignatureDef) colSchema() *openapi3.Schema {
 	}
 	return schema
 
-}
-
-func (t *TFSignatureDef) Schema() *openapi3.Schema {
-	// prefer the row format (https://www.tensorflow.org/tfx/serving/api_rest#specifying_input_tensors_in_row_format)
-	// when possible - it's more readable
-	if canHaveRowSchema(t.Inputs) {
-		return openapi3.NewObjectSchema().WithProperty("instances", t.rowSchema())
-	}
-
-	// column format (https://www.tensorflow.org/tfx/serving/api_rest#specifying_input_tensors_in_column_format)
-	colSchema := t.colSchema()
-	return openapi3.NewObjectSchema().WithProperty("inputs", colSchema)
 }
